requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-ninja==1.4.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
idna==3.11
imagekitio==4.2.0
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
pillow==12.0.0
praw==7.8.1
prawcore==2.4.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytube==15.0.0
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.9.0
yt-dlp==2025.10.22
